+++ Meeting-Management-System/src/main/java/Entity/Gender.java
@@ -10,7 +10,7 @@ import java.time.ZonedDateTime;
 public class Gender {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Short id;
+    private Long id;  // Changed from Short to Long
 
     @Column(name = "gender_alias", length = 10, nullable = false, unique = true)
     private String genderAlias;
