+++ Meeting-Management-System/src/main/java/Repository/EthnicCategoryRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EthnicCategoryRepo extends JpaRepository<EthnicCategory,Short> {
+public interface EthnicCategoryRepo extends JpaRepository<EthnicCategory,Long> {
 }
